pyproject.toml

[tool.poetry]
name = "binance_bot"
version = "0.1.0"
description = "A trading bot for Binance Futures Testnet"
authors = ["Yogesh Soni <yogesh.soni@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
python-binance = "^1.0.19"
click = "^8.1.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
